refactor shared InventoryCreationFragment functionalities with ShoppingCreationFragment into CreationFragment
